# implement cacheless lodash.get internally